pyproject.toml
[project]
name = "aegisota"
version = "0.1.0"
description = "Android OTA Upgrade Exception Injection and Multi-Device Verification Platform"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "sqlalchemy>=2.0.0",
    "typer>=0.9.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.24.0",
    "alembic>=1.18.4",
    "rapidfuzz>=3.0.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.0.280",
]

[project.scripts]
labctl = "app.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
no_implicit_optional = false

[[tool.mypy.overrides]]
module = "app.models.*"
disable_error_code = ["valid-type", "misc"]

[[tool.mypy.overrides]]
module = "tests.*"
disable_error_code = ["no-untyped-def", "arg-type"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "mypy>=1.20.1",
    "pre-commit>=4.5.1",
    "types-pyyaml>=6.0.12.20260408",
]
